Implementing set operations and comparisons Added different functionalities to compute and display the differences the 2 album sets (album_set1 and album_set2):
Added functionalites to compute and display the intersections of the 2 album sets:

Added functionalites to compute and display the union of the 2 album sets:

Implemented several checks to determine if one set is a superset or subset of the other:

Adressed Several QUizzes to demonstrate set operations:
-Considered the list  and set  :
Calculated and printed the sum of the elements  and  set .
This commit ensures all essential set operations are performed and their results printed. Also, included an illustrative quiz to contrast the summation of a list and a set, highlighting the elimination of duplicates in sets.